SystemUI: Allow Wi-Fi/cell tiles to co-exist with provider model
The dedicated Wi-Fi/cellular data QS tiles are partially working now
that we've exposed them in the list of tile options, but they're very
buggy and somewhat broken when the provider model (unified internet) is
enabled.

Allow the tiles to co-exist with InternetTile and the provider model
(including both settings_provider_model and combined signal icons) by
always including QS icons in Wi-Fi/data indicator callbacks, and add
a new flag to indicate whether they should be shown in the internet
tile.

Signed-off-by: Aston-Martinn <advaithbhat9@gmail.com>
Change-Id: I8affa7b3a74283b8becd0a785e707713b7d1e09d